Added PriceService and Implementation. Modified Repository and Model to be Integer instead of Long. Added DTO for clean code.

+++ src/main/java/com/inditex/codemain/model/Price.java
@@ -16,7 +16,7 @@ public class Price {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private Integer Id;
 
     private Integer brandId;
     private LocalDateTime startDate;
